requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
